+++ solutions_builder/modules/terraform_base/terraform/modules/firebase/variables.tf
@@ -24,7 +24,7 @@ variable "firestore_region" {
   type        = string
   description = "Firestore Region - must be app_engine region"
   # options for firestore: https://cloud.google.com/appengine/docs/locations
-  # {{gcp_region}} and europe-west1 must be us-central and europe-west for legacy reasons
+  # {{region}} and europe-west1 must be us-central and europe-west for legacy reasons
   default = "us-central"
 }